Add Q ModelLoader to load unredacted data when possible to avoid HEIC failures.
EXIF redaction in MediaStore on Q breaks HEIC/HEIF decoding. This class
does it's best to obtain unredacted file data on Q depending on the
state the hosting application is in with regards to storage.

It's not a complete fix. In particular applications that target Q, do
not opt in to legacy storage and do not have the ACCESS_MEDIA_LOCATION
permission will still be unable to decode HEIC/HEIF after this change.

PiperOrigin-RevId: 276114988